+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/desugar/StreamingCodeDesugar.java
@@ -1409,12 +1409,13 @@ public class StreamingCodeDesugar extends BLangNodeVisitor {
                     variableName = streamAliasMap.get(variableName);
                     ((BLangSimpleVarRef) ((BLangFieldBasedAccess) expr).expr).variableName.value = variableName;
                 }
-                streamEventParameter = createStringLiteral(expr.pos, (expr).toString());
-                exprs.set(i, streamEventParameter);
+                streamEventParameter = createStringLiteral(expr.pos, expr.toString());
+                exprs.set(i, desugar.addConversionExprIfRequired(streamEventParameter, symTable.anyType));
             } else if (expr.getKind() == NodeKind.LIST_CONSTRUCTOR_EXPR) {
                 BLangListConstructorExpr arrayLiteral = (BLangListConstructorExpr) expr;
                 convertFieldAccessArgsToStringLiteral(arrayLiteral.exprs);
             } else {
+                exprs.set(i, desugar.addConversionExprIfRequired(expr, symTable.anyType));
                 expr.typeChecked = false;
             }
         }
