+++ stdlib/http/src/test/java/org/ballerinalang/stdlib/services/nativeimpl/request/RequestNativeFunctionNegativeTest.java
@@ -84,9 +84,9 @@ public class RequestNativeFunctionNegativeTest {
         inRequest.set(REQUEST_ENTITY_FIELD, entity);
         BValue[] returnVals = BRunUtil.invoke(compileResult, "testGetJsonPayload", new Object[]{ inRequest });
         Assert.assertNotNull(returnVals[0]);
-        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"No payload\", "
-                + "cause:{ballerina/mime}ParsingEntityBodyFailed {cause:{ballerina/mime}NoContentError "
-                + "{message:\"Empty content\"}, message:\"Error occurred while extracting json data from entity\"}}");
+        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"No payload\", " +
+                "cause:{ballerina/mime}ParsingEntityBodyFailed {message:\"Error occurred while extracting json " +
+                "data from entity\", cause:{ballerina/mime}NoContentError {message:\"Empty content\"}}}");
     }
 
     @Test(description = "Test method with string payload")
@@ -123,9 +123,9 @@ public class RequestNativeFunctionNegativeTest {
         inRequest.set(REQUEST_ENTITY_FIELD, entity);
         BValue[] returnVals = BRunUtil.invoke(compileResult, "testGetTextPayload", new Object[]{ inRequest });
         Assert.assertTrue(returnVals[0].stringValue()
-                        .contains("{message:\"No payload\", cause:{ballerina/mime}ParsingEntityBodyFailed "
-                                + "{cause:{ballerina/mime}NoContentError {message:\"Empty content\"}, "
-                                + "message:\"Error occurred while extracting text data from entity\"}}"));
+                .contains("{message:\"No payload\", cause:{ballerina/mime}ParsingEntityBodyFailed " +
+                        "{message:\"Error occurred while extracting text data from entity\", " +
+                        "cause:{ballerina/mime}NoContentError {message:\"Empty content\"}}}"));
     }
 
     @Test
@@ -135,9 +135,9 @@ public class RequestNativeFunctionNegativeTest {
         TestEntityUtils.enrichTestEntityHeaders(entity, APPLICATION_XML);
         inRequest.set(REQUEST_ENTITY_FIELD, entity);
         BValue[] returnVals = BRunUtil.invoke(compileResult, "testGetXmlPayload", new Object[]{ inRequest });
-        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"No payload\", "
-                + "cause:{ballerina/mime}ParsingEntityBodyFailed {cause:{ballerina/mime}NoContentError "
-                + "{message:\"Empty content\"}, message:\"Error occurred while extracting xml data from entity\"}}");
+        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"No payload\", " +
+                "cause:{ballerina/mime}ParsingEntityBodyFailed {message:\"Error occurred while extracting xml data" +
+                " from entity\", cause:{ballerina/mime}NoContentError {message:\"Empty content\"}}}");
     }
 
     @Test
@@ -151,11 +151,11 @@ public class RequestNativeFunctionNegativeTest {
 
         BValue[] returnVals = BRunUtil.invoke(compileResult, "testGetXmlPayload", new Object[]{inRequest});
         Assert.assertNotNull(returnVals[0]);
-        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"Error occurred while "
-                + "retrieving the xml payload from the request\", cause:{ballerina/mime}ParsingEntityBodyFailed "
-                + "{cause:Unexpected character 'b' (code 98) in prolog; expected '<'" + StringUtil.NEWLINE
-                + " at [row,col {unknown-source}]: [1,1] {}, message:\"Error occurred while extracting xml "
-                + "data from entity\"}}");
+        Assert.assertEquals(((BError) returnVals[0]).getDetails().stringValue(), "{message:\"Error occurred while " +
+                "retrieving the xml payload from the request\", cause:{ballerina/mime}ParsingEntityBodyFailed " +
+                "{message:\"Error occurred while extracting xml data from entity\", cause:Unexpected " +
+                "character 'b' (code 98) in prolog; expected '<'" + StringUtil.NEWLINE +
+                " at [row,col {unknown-source}]: [1,1] {}}}");
     }
 
     @Test
