+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/statement/AbstractSQLStatement.java
@@ -85,6 +85,9 @@ import static org.ballerinax.jdbc.Constants.PARAMETER_VALUE_FIELD;
  * @since 1.0.0
  */
 public abstract class AbstractSQLStatement implements SQLStatement {
+
+    private static final BType SQL_PARAMETER_TYPE =
+            BallerinaValues.createRecordValue(Constants.JDBC_PACKAGE_ID, Constants.SQL_PARAMETER).getType();
     Calendar utcCalendar = Calendar.getInstance(TimeZone.getTimeZone(Constants.TIMEZONE_UTC));
     Strand strand;
 
@@ -93,7 +96,7 @@ public abstract class AbstractSQLStatement implements SQLStatement {
     }
 
     ArrayValue constructParameters(ArrayValue parameters) throws ApplicationException {
-        ArrayValue parametersNew = new ArrayValueImpl((BArrayType) parameters.getType());
+        ArrayValue parametersNew = new ArrayValueImpl(new BArrayType(SQL_PARAMETER_TYPE));
         int paramCount = parameters.size();
         for (int i = 0; i < paramCount; ++i) {
             Object value = parameters.get(i);
