+++ stdlib/messaging/kafka/src/main/java/org/ballerinalang/messaging/kafka/utils/KafkaUtils.java
@@ -46,7 +46,108 @@ import java.util.Objects;
 import java.util.Properties;
 
 import static org.ballerinalang.jvm.BallerinaValues.createRecord;
-import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.*;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_CONCURRENT_CONSUMERS;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_DECOUPLE_PROCESSING;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_OFFSET;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_PARTITION;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_POLLING_INTERVAL;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_POLLING_TIMEOUT;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_TOPIC;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ALIAS_TOPICS;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_AUTO_COMMIT_INTERVAL_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_AUTO_OFFSET_RESET_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_BOOTSTRAP_SERVERS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_CHECK_CRCS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_CLIENT_ID_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_CONFIG_FIELD_NAME;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_CONNECTIONS_MAX_IDLE_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_DEFAULT_API_TIMEOUT_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_ENABLE_AUTO_COMMIT_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_ERROR;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_EXCLUDE_INTERNAL_TOPICS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_FETCH_MAX_BYTES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_FETCH_MAX_WAIT_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_FETCH_MIN_BYTES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_GROUP_ID_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_HEARTBEAT_INTERVAL_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_INTERCEPTOR_CLASSES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_ISOLATION_LEVEL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_MAX_PARTITION_FETCH_BYTES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_MAX_POLL_INTERVAL_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_MAX_POLL_RECORDS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_METADATA_MAX_AGE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_METRICS_NUM_SAMPLES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_METRICS_RECORDING_LEVEL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_METRICS_SAMPLE_WINDOW_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_METRIC_REPORTER_CLASSES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_PARTITION_ASSIGNMENT_STRATEGY_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_RECEIVE_BUFFER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_RECONNECT_BACKOFF_MAX_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_RECONNECT_BACKOFF_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_RECORD_STRUCT_NAME;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_REQUEST_TIMEOUT_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_RETRY_BACKOFF_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_SEND_BUFFER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.CONSUMER_SESSION_TIMEOUT_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DEFAULT_KEY_DESERIALIZER;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DEFAULT_KEY_SERIALIZER;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DEFAULT_VALUE_DESERIALIZER;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DEFAULT_VALUE_SERIALIZER;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DETAIL_RECORD_NAME;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.DURATION_UNDEFINED_VALUE;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.ENABLED_PROTOCOLS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.KAFKA_PROTOCOL_PACKAGE_ID;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.KEYMANAGER_ALGORITHM_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.KEYSTORE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.KEYSTORE_TYPE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.LOCATION_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.NATIVE_PRODUCER;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.NATIVE_PRODUCER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.OFFSET_STRUCT_NAME;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PASSWORD_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_ACKS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_BATCH_SIZE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_BOOTSTRAP_SERVERS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_BUFFER_MEMORY_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_CLIENT_ID_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_COMPRESSION_TYPE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_CONNECTIONS_MAX_IDLE_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_ENABLE_IDEMPOTENCE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_INTERCEPTOR_CLASSES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_LINGER_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_MAX_BLOCK_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_MAX_IN_FLIGHT_REQUESTS_PER_CONNECTION;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_MAX_REQUEST_SIZE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_METADATA_MAX_AGE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_METRICS_NUM_SAMPLES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_METRICS_RECORDING_LEVEL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_METRICS_SAMPLE_WINDOW_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_METRIC_REPORTER_CLASSES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_PARTITIONER_CLASS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_RECEIVE_BUFFER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_RECONNECT_BACKOFF_MAX_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_RECONNECT_BACKOFF_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_REQUEST_TIMEOUT_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_RETRIES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_RETRY_BACKOFF_MS_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_SEND_BUFFER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_TRANSACTIONAL_ID_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PRODUCER_TRANSACTION_TIMEOUT_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PROPERTIES_ARRAY;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.PROTOCOL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SECURE_SOCKET;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SECURITY_PROTOCOL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_CIPHER_SUITES_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_KEY_PASSWORD_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_PROTOCOL_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_PROVIDER_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.SSL_SECURE_RANDOM_IMPLEMENTATION_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.TOPIC_PARTITION_STRUCT_NAME;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.TRUSTMANAGER_ALGORITHM_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.TRUSTSTORE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.TRUSTSTORE_TYPE_CONFIG;
+import static org.ballerinalang.messaging.kafka.utils.KafkaConstants.UNCHECKED;
 
 /**
  * Utility class for Kafka Connector Implementation.
