+++ stdlib/messaging/nats/src/main/java/org/ballerinalang/nats/basic/consumer/DefaultMessageHandler.java
@@ -20,7 +20,6 @@ package org.ballerinalang.nats.basic.consumer;
 
 import io.nats.client.Message;
 import io.nats.client.MessageHandler;
-
 import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.scheduling.Scheduler;
 import org.ballerinalang.jvm.services.ErrorHandlerUtils;
