+++ stdlib/mime/src/test/java/org/ballerinalang/stdlib/mime/MimeUtilityFunctionTest.java
@@ -21,7 +21,6 @@ package org.ballerinalang.stdlib.mime;
 import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.types.BMapType;
 import org.ballerinalang.jvm.types.BTypes;
-import org.ballerinalang.jvm.values.ArrayValue;
 import org.ballerinalang.jvm.values.ArrayValueImpl;
 import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.MapValueImpl;
