+++ stdlib/socket/src/main/java/org/ballerinalang/stdlib/socket/tcp/SelectorManager.java
@@ -56,8 +56,8 @@ import java.util.concurrent.ThreadFactory;
 
 import static java.nio.channels.SelectionKey.OP_READ;
 import static org.ballerinalang.stdlib.socket.SocketConstants.DEFAULT_EXPECTED_READ_LENGTH;
-import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE_ID;
 import static org.ballerinalang.stdlib.socket.SocketConstants.ErrorCode.ReadTimedOutError;
+import static org.ballerinalang.stdlib.socket.SocketConstants.SOCKET_PACKAGE_ID;
 
 /**
  * This will manage the Selector instance and handle the accept, read and write operations.
