+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/service/grpc/sample/UnaryBlockingArrayValueTestCase.java
@@ -21,6 +21,7 @@ package org.ballerinalang.test.service.grpc.sample;
 import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.types.BPackage;
 import org.ballerinalang.jvm.values.ArrayValue;
+import org.ballerinalang.jvm.values.ArrayValueImpl;
 import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.model.values.BBoolean;
 import org.ballerinalang.model.values.BFloat;
