+++ tests/jballerina-integration-test/src/test/java/org/ballerinalang/test/service/grpc/tool/ProtoMessageTestCase.java
@@ -22,7 +22,7 @@ import com.google.protobuf.Descriptors;
 import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.types.BPackage;
 import org.ballerinalang.jvm.types.BRecordType;
-import org.ballerinalang.jvm.values.ArrayValue;
+import org.ballerinalang.jvm.values.ArrayValueImpl;
 import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.net.grpc.Message;
 import org.ballerinalang.net.grpc.MessageParser;
